requirements.txt
flask-appbuilder
apscheduler
python-slugify
opencv-contrib-python
